requirements.txt
# FastAPI & server
fastapi==0.115.2
uvicorn[standard]==0.30.6
httpx==0.27.2
cryptography==43.0.3

# RAG
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
sentence-transformers==3.0.1
numpy==1.26.4
pydantic==2.9.2
pydantic-settings==2.6.1

# Document parsing (multi-format support)
pdfplumber==0.11.4
python-docx==1.1.2
chardet==5.2.0